requirements.txt
flask
ariadne
pipflask-cors